# perf(behavior_velocity_planner): replace getSortedLaneIdsFromPath