@Enigamict 2025-02-12
@veqcc @soblin 
Thank you for the review. I now understand the behavior of getSortedLaneIdsFromPath and appreciate the explanation about lane_ids. I realize that the current modification doesn't consider the order, so it needs to be adjusted. I'll make the necessary corrections based on your feedback.