@Enigamict 2025-02-25
@veqcc @soblin
We apologize for any inconvenience caused. After trying various methods, we have seen no improvement in performance, and the traditional method remains the best at this time.

Thank you very much for your review and feedback. We will now close this pull request.